requirements.txt
flask==3.0.0
sqlalchemy==2.0.23
argon2-cffi==23.1.0
cryptography==41.0.7
pyjwt==2.8.0
pyotp==2.9.0
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0
user-agents==2.2.0